pyproject.toml
[project]
name = "twilio-morning-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "feedparser>=6.0.11",
    "flask[async]>=3.1.1",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "twilio>=9.7.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
